requirements.txt
discord
python-dotenv
vaderSentiment
